Move path control to separae source file
