+++ CMakeLists.txt
@@ -90,6 +90,7 @@ set(PICOQUIC_LIBRARY_FILES
     picoquic/newreno.c
     picoquic/pacing.c
     picoquic/packet.c
+    picoquic/paths.c
     picoquic/performance_log.c
     picoquic/picohash.c
     picoquic/picoquic_lb.c
