+++ picoquic/picoquic.vcxproj.filters
@@ -144,6 +144,9 @@
     <ClCompile Include="register_all_cc_algorithms.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="paths.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="picoquic.h">
